增加播放方法，可修改句子顺序

+++ app/src/pages/word/sentenceList/Mp3Time/index.tsx
@@ -18,7 +18,7 @@ const Mp3Time: React.FC = (props) => {
             const timeArr = value.split(",")
             if (timeArr.length == 2) {
                 const startTime = parseInt(timeArr[0])
-                const duration = parseInt(timeArr[1])
+                const duration = parseFloat(timeArr[1])
                 setStartTime(startTime)
                 setEndTime(startTime + duration)
             }
